make icon links a component that shows three links instead of a component that displays one and needs to be wrapped three times whenever we use it
