--- less/components/icon-link.less
@@ -1,25 +1,0 @@
-.icon-link {
-  text-align: center;
-  img {
-    display: block;
-    margin: 0 auto;
-    transition: all 0.18s ease-in;
-    &:hover {
-      transform: scale(1.1);
-    }
-  }
-  h3 {
-    margin-bottom: 0.5rem;
-  }
-  .subhead {
-    font-style: italic;
-    font-size: 1.6rem;
-  }
-}
-
-.icon-links {
-  margin-bottom: -2rem;
-  .icon-link-container {
-    padding-bottom: 2rem;
-  }
-}
