+++ less/components/index.less
@@ -3,7 +3,7 @@
 @import "sidebar";
 @import "footer";
 @import "hero-unit";
-@import "icon-link";
+@import "icon-links";
 @import "imagetag";
 @import "modal";
 @import "map";
